refactor(reposicao): extrai AdminReposicaoOportunidades god-component (1093→752)
5º refactor da série de god-components (§10). Extrai types/helpers e subcomponentes
de AdminReposicaoOportunidades.tsx para src/components/reposicao/oportunidades/.
Extração pura, behavior-preserving:

- types.ts       — Cenario, AumentoRef, Oportunidade, OrdemKey
- shared.tsx     — EMPRESA, ALL, CENARIOS, formatBRL, formatNumber, formatDate,
                   formatDateLong, diasEntre, cenarioIcon, cenarioLabel,
                   descontoBadgeClass, diasBadge
- components.tsx — EstadoVazio, DrawerConteudo, Linha

A página vira orquestrador (752 linhas). Imports órfãos podados (arquivo está no
tsconfig.strict.json → noUnusedLocals pegou). Revisado via codex review (extração
fiel, strict+app passam). tsc + strict + 396/396 testes + lint 0.
